examples/pydantic-ai-research/pyproject.toml
[project]
name = "majordomo-pydantic-ai-example"
version = "0.1.0"
description = "Pydantic AI agent examples with Majordomo Gateway integration"
requires-python = ">=3.12"
dependencies = [
    "majordomo-frameworks[pydantic-ai]",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.9.0",
]

[tool.uv.sources]
majordomo-frameworks = { path = "../.." }
